+++ src/main/java/com/taisheng/now/bussiness/bean/result/market/OrderGoodsBean.java
@@ -1,10 +1,12 @@
 package com.taisheng.now.bussiness.bean.result.market;
 
+import java.math.BigDecimal;
+
 public class OrderGoodsBean {
 
     public String picUrl;
     public int number;
-    public int price;
+    public BigDecimal price;
     public String name;
     public String id;
     public String goodsId;
